+++ deltafi-ui/src/components/errors/SummaryByMessagePanel.vue
@@ -158,7 +158,7 @@ const filterSelectedDids = computed(() => {
 
 const fetchErrorsMessages = async () => {
   getPersistedParams();
-  let ingressFlowName = props.ingressFlowName != null ? props.ingressFlowName.name : null;
+  let ingressFlowName = props.ingressFlowName != null ? props.ingressFlowName : null;
   let showAcknowled = props.awknowledged ? null : false;
   loading.value = true;
   await getErrorsByMessage(showAcknowled, offset.value, perPage.value, sortField.value, sortDirection.value, ingressFlowName);
@@ -244,6 +244,4 @@ const setPersistedParams = () => {
 };
 </script>
 
-<style lang="scss">
-
-</style>
+<style lang="scss"></style>